- No idea why I have to manually commit this ;-)

+++ plc4j/spi/src/main/java/org/apache/plc4x/java/spi/generation/ReadBuffer.java
@@ -42,7 +42,7 @@ public class ReadBuffer {
         ArrayByteInput abi = new ArrayByteInput(input);
         this.bi = new MyDefaultBitInput(abi);
         this.littleEndian = littleEndian;
-        this.totalBytes = (long) input.length;
+        this.totalBytes = input.length;
     }
 
     public int getPos() {
